ln hooks + refacto
