+++ packages/afk_nostr_sdk/src/hooks/index.ts
@@ -48,3 +48,4 @@ export {useSendZap} from './zap/useZap';
 export {useCashu} from "./cashu/useCashu"
 export {useConnectNWC} from './zap/useZap';
 export {useSendZapNote} from './zap/useZap';
+export {useLN} from "./ln";
